+++ packages/@enotion/components/src/index.ts
@@ -6,3 +6,4 @@ export * from "./Input.js";
 export * from "./Link.js";
 export * from "./Select.js";
 export * from "./Card.js";
+export * from "./Search.js";